split logic: separate main page adn quiz page
