+++ app/templates/index.html
@@ -3,34 +3,16 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Опросник</title>
+    <title>Главная страница</title>
     <link rel="stylesheet" href="../static/css/styles.css">
 </head>
 <body>
 
-    <header>
-        <h1>Опросник</h1>
-        <div class="progress-bar">
-            <div class="progress"></div>
-        </div>
-        <span id="step-counter">1/5</span>
-    </header>
-
-    <!-- Экран выбора действия -->
     <main id="start-screen">
         <button id="startQuizBtn">Пройти опросник</button>
         <button id="loginBtn">Авторизоваться</button>
     </main>
 
-    <!-- Опросник (скрыт по умолчанию) -->
-    <main id="quiz-container" style="display: none;">
-        <div id="question-container"></div>
-        <div class="buttons">
-            <button id="prevBtn" disabled>Назад</button>
-            <button id="nextBtn" disabled>Далее</button>
-        </div>
-    </main>
-
-    <script type="module" src="../static/js/main.js"></script>
+    <script src="../static/js/main.js"></script>
 </body>
 </html>
